Python Studying Aid/requirements.txt
#Need your own 
    #1. Spotify Client ID, Spotify Client Secret 
    #2. Unsplash Access Key, Unsplash Secret Key, Unsplash Application ID 
#Spotify API credentials in spotify_controller.py
#Unplash API credentials in base_page.py

annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
charset-normalizer==3.4.6
colorama==0.4.6
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
keyboard==0.13.5
pillow==12.1.1
pydantic==2.12.5
pydantic_core==2.41.5
pypiwin32==223
pystray==0.19.5
pywin32==311
redis==7.3.0
requests==2.32.5
six==1.17.0
sniffio==1.3.1
spotipy==2.26.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3